fix(seguridad): escapar la constante del limite y traer los sniffers de PCP al repo
La CI (Plugin Check) rechazo la 1.6.0 por un printf con la constante del limite sin
escapar. El gate local no lo cogia porque en el repo no habia phpcs, asi que ese paso
solo se veia al subir.

- Se castea a int lo que va dentro del printf.
- phpcs + WordPress Coding Standards como dependencia de desarrollo, y script
  `composer test:security` con los sniffers que mira PCP (escapado, nonce y saneado),
  incluido en `composer test`.

+++ includes/class-admin.php
@@ -819,7 +819,7 @@ class Admin
                             printf(
                                 /* translators: %d: número de cuentas de esa red */
                                 esc_html(_n('%d cuenta', '%d cuentas', count($own), 'convoca-publisher')),
-                                count($own)
+                                (int) count($own)
                             );
                         ?>
                         </span>
@@ -871,7 +871,7 @@ class Admin
                         printf(
                             /* translators: %d: límite de cuentas por red */
                             esc_html__('Límite alcanzado: %d cuentas por red. Borra una para añadir otra.', 'convoca-publisher'),
-                            Profile_Store::LIMIT_PER_NETWORK
+                            (int) Profile_Store::LIMIT_PER_NETWORK
                         );
                     ?>
                     </p>
